requirements.txt
fastapi
uvicorn
pymongo
pydantic
pydantic-settings
pandas
numpy==1.26.4
scikit-learn
python-dotenv
streamlit